[chore #161839122] rename orders folder and functions as recommended by LFA
