+++ .coverage
@@ -1,1 +1,1 @@
-!coverage.py: This is a private format, don't read it directly!{"lines":{"D:\\SendIT-API\\api\\__init__.py":[1],"D:\\SendIT-API\\api\\test\\__init__.py":[1],"D:\\SendIT-API\\api\\test\\test_cancel_orders.py":[1,2,3,5,6,7,38,43,48,8,9,14,15,16,17,18,19,20,21,22,26,27,28,29,30,31,32,33,34,39,40,44,45,49,50],"D:\\SendIT-API\\api\\version1\\__init__.py":[1],"D:\\SendIT-API\\api\\version1\\all_orders\\__init__.py":[1,2,4],"D:\\SendIT-API\\api\\version1\\all_orders\\views.py":[1,2,3,5,6,8,15,22,29,36,43,45,46,17,18,24,25,31,32,38,39],"D:\\SendIT-API\\api\\version1\\all_orders\\models.py":[1,2,6,7,8,9,10,11,12,13,14,17,18,19,20,21,22,23,24,25,29,30,35,62,68,75,82,31,32,83,84,91,85,86,89,90,38,39,40,41,42,43,44,45,46,49,51,55,59,52,53,63,65,69,70,71,72,76,77,78,79],"D:\\SendIT-API\\api\\test\\test_create_orders.py":[1,2,3,5,6,7,33,37,8,9,12,13,14,15,16,17,18,19,23,24,25,26,27,28,29,30,34,35,38,39],"D:\\SendIT-API\\api\\test\\test_get_orders.py":[1,2,3,5,6,7,39,44,49,54,59,8,9,15,16,17,18,19,20,21,22,23,26,27,28,29,30,31,32,33,34,40,41,45,46,50,51,55,56,60,61]}}
+!coverage.py: This is a private format, don't read it directly!{"lines":{"D:\\SendIT-API\\api\\__init__.py":[1],"D:\\SendIT-API\\api\\test\\__init__.py":[1],"D:\\SendIT-API\\api\\test\\test_cancel_orders.py":[1,2,3,5,6,7,38,43,48,8,9,14,15,16,17,18,19,20,21,22,26,27,28,29,30,31,32,33,34,39,40,44,45,49,50],"D:\\SendIT-API\\api\\version1\\__init__.py":[1],"D:\\SendIT-API\\api\\version1\\orders\\__init__.py":[1,2,4],"D:\\SendIT-API\\api\\version1\\orders\\views.py":[1,2,3,5,6,8,15,22,29,36,43,45,46,17,18,24,25,38,39,31,32],"D:\\SendIT-API\\api\\version1\\orders\\models.py":[1,2,6,7,8,9,10,11,12,13,14,17,18,19,20,21,22,23,24,25,29,30,35,62,68,75,82,31,32,83,84,91,85,86,89,90,38,39,40,41,42,43,44,45,46,49,51,55,59,52,53,63,65,76,77,78,79,69,70,71,72],"D:\\SendIT-API\\api\\test\\test_create_orders.py":[1,2,3,5,6,7,33,37,8,9,12,13,14,15,16,17,18,19,23,24,25,26,27,28,29,30,34,35,38,39],"D:\\SendIT-API\\api\\test\\test_get_orders.py":[1,2,3,5,6,7,39,44,49,54,59,8,9,15,16,17,18,19,20,21,22,23,26,27,28,29,30,31,32,33,34,40,41,55,56,60,61,45,46,50,51]}}
