+++ api/test/test_cancel_orders.py
@@ -1,6 +1,6 @@
 import unittest
 import json
-from api.version1.all_orders import app
+from api.version1.orders import app
 
 class TestOrders(unittest.TestCase):
 	"""This class tests the fetching of orders"""
@@ -35,16 +35,16 @@ class TestOrders(unittest.TestCase):
 			} 
 	
 	"""Test cancel order for ID that exists and status ==delivered"""
-	def test_cancel_orders_1(self):
+	def test_cancel_order_1(self):
 		result = self.client.put('/api/v1/cancel/1', data = json.dumps(self.order1), content_type='application/json')
 		self.assertEqual(result.status_code, 200)
 
 	"""Test cancel order for ID that exists and status !=delivered """
-	def test_cancel_orders_2(self):
+	def test_cancel_order_2(self):
 		result = self.client.put('/api/v1/cancel/2', data = json.dumps(self.order2), content_type='application/json')
 		self.assertEqual(result.status_code, 200)
 
 	"""Test cancel order for ID that does not exist """
-	def test_cancel_orders_3(self):
+	def test_cancel_order_3(self):
 		result = self.client.put('/api/v1/cancel/3', data = json.dumps(self.order2), content_type='application/json')
 		self.assertEqual(result.status_code, 200)
