+++ api/version1/orders/__init__.py
@@ -1,4 +1,4 @@
 from flask import Flask
 app = Flask(__name__)
 
-import api.version1.all_orders.views        
+import api.version1.orders.views
