+++ api/version1/orders/models.py
@@ -1,4 +1,4 @@
-from api.version1.all_orders import app
+from api.version1.orders import app
 from flask import request, jsonify
 
 orders_list = [
@@ -64,22 +64,22 @@ class SendItOrders():
 			return {"message" : "No orders in the system"}
 		return orders_list
 
-	"""fetch all orders in the system by ID"""
-	def all_orders_by_parcelID(self, parcelID):
+	"""fetch specific order in the system by ID"""
+	def order_by_ID(self, parcelID):
 		order = [order for order in orders_list if order['parcelID'] == parcelID]
 		if order:
 			return order
 		return {"message" : "No orders with that ID"}
 
-	"""fetch all orders in the system by email"""
+	"""fetch all orders in the system by specific email"""
 	def all_orders_by_sender_email(self, sender_email):
 		order = [order for order in orders_list if order['sender_email'] == sender_email]
 		if order:
 			return order
 		return {"message" : "No orders made from that email address"}
 
-	"""cancel orders"""
-	def cancel_orders(self, parcelID):
+	"""cancel order specified by parcelID"""
+	def cancel_order(self, parcelID):
 		order = [order for order in orders_list if (order['parcelID'] == parcelID and order['status'] !='Delivered')]
 		if order:
 			order[0]["status"] = request.json["status"]
